+++ doc/source/changelog.rst
@@ -16,6 +16,7 @@ master: 1.7.1
 * Expose protected members in Python derived classes
 * Fix by-ptr return of objects in Python derived classes
 * Fix for passing numpy boolean array through ``bool*``
+* Fix assignment to ``const char*`` data members
 
 
 2020-04-27: 1.7.0